backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
autobahn==19.11.2
Automat==25.4.16
build==1.3.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
constantly==23.10.4
cryptography==46.0.3
fastapi==0.127.0
h11==0.16.0
httptools==0.7.1
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
kiteconnect==5.0.1
numpy==2.4.0
packaging==25.0
pandas==2.3.3
patsy==1.0.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pyOpenSSL==25.3.0
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
scipy==1.16.3
service-identity==24.2.0
six==1.17.0
starlette==0.50.0
statsmodels==0.14.6
TA-Lib==0.6.8
Twisted==25.5.0
txaio==25.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
uvicorn==0.40.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
zope.interface==8.1.1
